added timeout duration
Turns out that Applescripts time out, by default, after two minutes. So if the person running this script doesn't reply within that two minutes, the script times out in the background and the dialog just sits there. To make matters worse, if this is a part of a chain of policies then the following policies start to execute - if any of those subsequent policies rely on the answer to this dialog then it causes a cascade of errors.